Move isPlaying and duration to generic hook library to allow for multiple projectCards
